+++ minecraft/archCommon/src/main/java/io/github/kosmx/emotes/arch/screen/widget/LegacyChooseWidget.java
@@ -3,7 +3,7 @@ package io.github.kosmx.emotes.arch.screen.widget;
 
 import com.mojang.blaze3d.systems.RenderSystem;
 import dev.kosmx.playerAnim.core.util.MathHelper;
-import io.github.kosmx.emotes.arch.executor.Defaults;
+import io.github.kosmx.emotes.PlatformTools;
 import io.github.kosmx.emotes.executor.EmoteInstance;
 import io.github.kosmx.emotes.main.EmoteHolder;
 import io.github.kosmx.emotes.main.config.ClientConfig;
@@ -22,7 +22,7 @@ public class LegacyChooseWidget implements IChooseWheel {
     //protected final FastChooseElement[] elements = new FastChooseElement[8];
     protected final ArrayList<FastChooseElement> elements = new ArrayList<>();
     private boolean hovered;
-    private final ResourceLocation TEXTURE = ((ClientConfig) EmoteInstance.config).dark.get() ? Defaults.newIdentifier("textures/gui/fastchoose_dark.png") : Defaults.newIdentifier("textures/gui/fastchoose_light.png");
+    private final ResourceLocation TEXTURE = ((ClientConfig) EmoteInstance.config).dark.get() ? PlatformTools.newIdentifier("textures/gui/fastchoose_dark.png") : PlatformTools.newIdentifier("textures/gui/fastchoose_light.png");
 
     private final AbstractFastChooseWidget widget;
