+++ minecraft/archCommon/src/main/java/io/github/kosmx/emotes/arch/screen/widget/ModernChooseWheel.java
@@ -1,8 +1,8 @@
 package io.github.kosmx.emotes.arch.screen.widget;
 
 import com.mojang.blaze3d.systems.RenderSystem;
+import io.github.kosmx.emotes.PlatformTools;
 import io.github.kosmx.emotes.executor.EmoteInstance;
-import io.github.kosmx.emotes.inline.TmpGetters;
 import io.github.kosmx.emotes.main.EmoteHolder;
 import io.github.kosmx.emotes.main.config.ClientConfig;
 import net.minecraft.client.Minecraft;
@@ -28,7 +28,7 @@ public class ModernChooseWheel implements IChooseWheel {
     //protected final FastChooseElement[] elements = new FastChooseElement[8];
     protected final ArrayList<FastChooseElement> elements = new ArrayList<>();
     private boolean hovered;
-    private final ResourceLocation TEXTURE = ((ClientConfig) EmoteInstance.config).dark.get() ? TmpGetters.getDefaults().newIdentifier("textures/gui/fastchoose_dark_new.png") : TmpGetters.getDefaults().newIdentifier("textures/gui/fastchoose_light_new.png");
+    private final ResourceLocation TEXTURE = ((ClientConfig) EmoteInstance.config).dark.get() ? PlatformTools.newIdentifier("textures/gui/fastchoose_dark_new.png") : PlatformTools.newIdentifier("textures/gui/fastchoose_light_new.png");
 
     private final AbstractFastChooseWidget widget;
