+++ minecraft/archCommon/src/main/java/io/github/kosmx/emotes/inline/TmpGetters.java
@@ -6,10 +6,6 @@ import io.github.kosmx.emotes.arch.executor.Defaults;
 public class TmpGetters {
 
 
-    public static Defaults getDefaults() {
-        return new Defaults();
-    }
-
     public static ClientMethods getClientMethods() {
         return new ClientMethods();
     }
