+++ minecraft/archCommon/src/main/java/io/github/kosmx/emotes/main/network/ClientPacketManager.java
@@ -1,11 +1,11 @@
 package io.github.kosmx.emotes.main.network;
 
+import io.github.kosmx.emotes.PlatformTools;
 import io.github.kosmx.emotes.api.proxy.EmotesProxyManager;
 import io.github.kosmx.emotes.api.proxy.INetworkInstance;
 import io.github.kosmx.emotes.common.network.EmotePacket;
 import io.github.kosmx.emotes.common.network.objects.NetData;
 import io.github.kosmx.emotes.executor.EmoteInstance;
-import io.github.kosmx.emotes.inline.TmpGetters;
 import io.github.kosmx.emotes.main.EmoteHolder;
 
 import java.io.IOException;
@@ -19,7 +19,7 @@ import java.util.logging.Level;
  */
 public final class ClientPacketManager extends EmotesProxyManager {
 
-    private static final INetworkInstance defaultNetwork = TmpGetters.getClientMethods().getServerNetworkController();
+    private static final INetworkInstance defaultNetwork = PlatformTools.getClientNetworkController();
     //that casting should always work
 
     public static void init(){
